requirements.txt
numpy>=1.19.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
gym>=0.21.0
stable-baselines3>=1.5.0
ta>=0.7.0
torch>=1.10.0
scikit-learn>=1.0.0
pyyaml>=6.0
tqdm==4.66.2